Phase 2 Wave 4: split Quota, Social, Contribution, and Publication slices out of store.ts
Moves QuotaStore/GlobalQuotaStore, SocialStore, ContributionStore, and
PublicationStore implementations out of the monolithic store.ts into
apps/api/src/store/slices/*.ts, following the delegation pattern established
by prior batches (#961, #963, #964).

- quota.ts / quota-global.ts: QuotaStore split into per-user quota (usage,
  creation limits, public-play/featured-pool config) and GlobalQuotaStore
  (the nine shared daily-allowance counters), since the combined file
  exceeded the 500-line new-file cap. InMemoryQuotaStore takes a getUser
  callback rather than reaching into the identity slice directly.
- social.ts: SocialStore (votes, follows, player feedback). votes/follows/
  playerFeedback stay non-private -- PublicationStore.listGameSlugs reaches
  across them (documented exception).
- contribution.ts: ContributionStore (suggestions, proposals, contribution
  settings, contributor blocks, legacy-suggestion purge). suggestions and
  gameAutonomy stay non-private -- deleteAccountIdentity reaches across them.
- publication.ts: PublicationStore (publish/takedown/archive/health-check).
  listGameSlugs stays directly on the composition root as a new minimal
  GameSlugsStore interface (same pattern as AccountErasureStore) since its
  in-memory implementation unions keys across Social, Contribution, and
  Review's Maps -- a cross-cutting read a delegate-only slice must not own.

All five new files pass the comment-prose and module-size checks at true
baseline 0, per the lesson from PR #964. store.ts shrinks from 5112 to 4081
lines.
